split ocean-api rawtx 1 test 1 file
